scripts/cloud_functions/ephemeral_cleanup/requirements.txt
google-cloud-compute==1.13.0
google-cloud-container==11.9.0
google-cloud-run==0.10.0
google-cloud-storage==2.10.0
google-cloud-logging==3.5.0
functions-framework==3.3.0
